SVY-10550 support servoy.max.webforms.loaded in ng solutions
